server/auth: implement access control and peer account creation on host
STATE: At this point the server receives an authorization request from a
peer client to a host client. The server installs the requested access
on the host client.

The peer client can request credentials to the host client. In that case
the server creates new PSK and installs them together with the access
rights on the host client.

Installing host client account on peer client is still missing.